#27 extract common bits to the 'common' project
